+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [NEXT RELEASE]
 
 ### Added
+- Data Plane terraforming scripts migration from BitWarden to Parameter Store
 ### Changed
 ### Deprecated
 ### Removed